post first commit

+++ Back/src/main/java/com/example/stoic/Comment/Repo/CommentRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepo extends JpaRepository<Comment, Long> {
+public interface CommentRepo extends JpaRepository<Comment, Integer> {
     // Add custom query methods if needed
 }
